+++ Keyboards/LanguageKeyboards/Spanish/ESInterfaceVariables.swift
@@ -20,90 +20,8 @@
 import UIKit
 
 public enum SpanishKeyboardConstants {
-  // iPhone keyboard layouts.
-  static let letterKeysPhone = [
-    ["q", "w", "e", "r", "t", "y", "u", "i", "o", "p"],
-    ["a", "s", "d", "f", "g", "h", "j", "k", "l", "ñ"],
-    ["shift", "z", "x", "c", "v", "b", "n", "m", "delete"],
-    ["123", "selectKeyboard", "space", "return"] // "undo"
-  ]
-
-  static let letterKeysPhoneDisableAccents = [
-    ["q", "w", "e", "r", "t", "y", "u", "i", "o", "p"],
-    ["a", "s", "d", "f", "g", "h", "j", "k", "l"],
-    ["shift", "z", "x", "c", "v", "b", "n", "m", "delete"],
-    ["123", "selectKeyboard", "space", "return"] // "undo"
-  ]
-
-  static let numberKeysPhone = [
-    ["1", "2", "3", "4", "5", "6", "7", "8", "9", "0"],
-    ["-", "/", ":", ";", "(", ")", "$", "&", "@", "\""],
-    ["#+=", ".", ",", "?", "!", "'", "delete"],
-    ["ABC", "selectKeyboard", "space", "return"] // "undo"
-  ]
-
-  static let symbolKeysPhone = [
-    ["[", "]", "{", "}", "#", "%", "^", "*", "+", "="],
-    ["_", "\\", "|", "~", "<", ">", "€", "£", "¥", "·"],
-    ["123", ".", ",", "?", "!", "'", "delete"],
-    ["ABC", "selectKeyboard", "space", "return"] // "undo"
-  ]
-
-  // iPad keyboard layouts.
-  static let letterKeysPad = [
-    ["1", "2", "3", "4", "5", "6", "7", "8", "9", "0", "-", "=", "+"],
-    ["q", "w", "e", "r", "t", "y", "u", "i", "o", "p", "delete"],
-    ["a", "s", "d", "f", "g", "h", "j", "k", "l", "ñ", "return"],
-    ["shift", "z", "x", "c", "v", "b", "n", "m", ",", ".", "shift"],
-    ["selectKeyboard", ".?123", "space", ".?123", "hideKeyboard"] // "undo"
-  ]
-
-  static let letterKeysPadDisableAccents = [
-    ["1", "2", "3", "4", "5", "6", "7", "8", "9", "0", "-", "=", "+"],
-    ["q", "w", "e", "r", "t", "y", "u", "i", "o", "p", "delete"],
-    ["a", "s", "d", "f", "g", "h", "j", "k", "l", "return"],
-    ["shift", "z", "x", "c", "v", "b", "n", "m", ",", ".", "shift"],
-    ["selectKeyboard", ".?123", "space", ".?123", "hideKeyboard"] // "undo"
-  ]
-
-  static let numberKeysPad = [
-    ["1", "2", "3", "4", "5", "6", "7", "8", "9", "0", "delete"],
-    ["@", "#", "$", "&", "*", "(", ")", "'", "\"", "+", "return"],
-    ["#+=", "%", "_", "-", "=", "/", ";", ":", ",", ".", "#+="],
-    ["selectKeyboard", "ABC", "space", "ABC", "hideKeyboard"] // "undo"
-  ]
-
-  static let symbolKeysPad = [
-    ["1", "2", "3", "4", "5", "6", "7", "8", "9", "0", "delete"],
-    ["€", "£", "¥", "^", "[", "]", "{", "}", "ᵒ", "ᵃ", "return"],
-    ["123", "§", "|", "~", "¶", "\\", "<", ">", "¡", "¿", "123"],
-    ["selectKeyboard", "ABC", "space", "ABC", "hideKeyboard"] // "undo"
-  ]
-
-  // Expanded iPad keyboard layouts for wider devices.
-  static let letterKeysPadExpanded = [
-    ["|", "1", "2", "3", "4", "5", "6", "7", "8", "9", "0", "?", "¿", "delete"],
-    [SpecialKeys.indent, "q", "w", "e", "r", "t", "y", "u", "i", "o", "p", "´", "+", "*"],
-    [SpecialKeys.capsLock, "a", "s", "d", "f", "g", "h", "j", "k", "l", "ñ", "{", "}", "return"],
-    ["shift", "'", "z", "x", "c", "v", "b", "n", "m", ",", ".", "-", "shift"],
-    ["selectKeyboard", ".?123", "space", ".?123", "hideKeyboard"] // "microphone", "scribble"
-  ]
-
-  static let letterKeysPadExpandedDisableAccents = [
-    ["|", "1", "2", "3", "4", "5", "6", "7", "8", "9", "0", "?", "¿", "delete"],
-    [SpecialKeys.indent, "q", "w", "e", "r", "t", "y", "u", "i", "o", "p", "´", "+", "*"],
-    [SpecialKeys.capsLock, "a", "s", "d", "f", "g", "h", "j", "k", "l", "~", "{", "}", "return"],
-    ["shift", "'", "z", "x", "c", "v", "b", "n", "m", ",", ".", "-", "shift"],
-    ["selectKeyboard", ".?123", "space", ".?123", "hideKeyboard"] // "microphone", "scribble"
-  ]
-
-  static let symbolKeysPadExpanded = [
-    ["`", "1", "2", "3", "4", "5", "6", "7", "8", "9", "0", "<", ">", "delete"],
-    [SpecialKeys.indent, "(", ")", "{", "}", "#", "%", "^", "*", "+", "=", "\\", "|", "§"],
-    [SpecialKeys.capsLock, "—", "/", ":", ";", "&", "@", "$", "£", "¥", "~", "[", "]", "return"], // "undo"
-    ["shift", "…", "?", "!", "≠", "'", "\"", "_", "€", ",", ".", "-", "shift"], // "redo"
-    ["selectKeyboard", "ABC", "space", "ABC", "hideKeyboard"] // "microphone", "scribble"
-  ]
+  static let defaultCurrencyKey = "$"
+  static let currencyKeys = ["$", "€", "£", "¥"]
 
   // Alternate key vars.
   static let keysWithAlternates = ["a", "e", "i", "o", "u", "c", "d", "n", "s"]
@@ -122,18 +40,161 @@ public enum SpanishKeyboardConstants {
   static let sAlternateKeys = ["š"]
 }
 
+struct SpanishKeyboardProvider: KeyboardProviderProtocol, KeyboardProviderDisableAccentsProtocol {
+  // iPhone keyboard layouts.
+  static func genPhoneLetterKeys() -> [[String]] {
+    return KeyboardBuilder()
+      .addRow(["q", "w", "e", "r", "t", "y", "u", "i", "o", "p"])
+      .addRow(["a", "s", "d", "f", "g", "h", "j", "k", "l", "ñ"])
+      .addRow(["shift", "z", "x", "c", "v", "b", "n", "m", "delete"])
+      .addRow(["123", "selectKeyboard", "space", "return"]) // "undo"
+      .build()
+  }
+
+  static func genPhoneDisableAccentsLetterKeys() -> [[String]] {
+    return KeyboardBuilder()
+      .addRow(["q", "w", "e", "r", "t", "y", "u", "i", "o", "p"])
+      .addRow(["a", "s", "d", "f", "g", "h", "j", "k", "l"])
+      .addRow(["shift", "z", "x", "c", "v", "b", "n", "m", "delete"])
+      .addRow(["123", "selectKeyboard", "space", "return"]) // "undo"
+      .build()
+  }
+
+  static func genPhoneNumberKeys(currencyKey: String) -> [[String]] {
+    return KeyboardBuilder()
+      .addRow(["1", "2", "3", "4", "5", "6", "7", "8", "9", "0"])
+      .addRow(["-", "/", ":", ";", "(", ")", "$", "&", "@", "\""])
+      .addRow(["#+=", ".", ",", "?", "!", "'", "delete"])
+      .addRow(["ABC", "selectKeyboard", "space", "return"]) // "undo"
+      .replaceKey(row: 1, column: 6, to: currencyKey)
+      .build()
+  }
+
+  static func genPhoneSymbolKeys(currencyKeys: [String]) -> [[String]] {
+    let keyboardBuilder = KeyboardBuilder()
+      .addRow(["[", "]", "{", "}", "#", "%", "^", "*", "+", "="])
+      .addRow(["_", "\\", "|", "~", "<", ">", "€", "£", "¥", "·"])
+      .addRow(["123", ".", ",", "?", "!", "'", "delete"])
+      .addRow(["ABC", "selectKeyboard", "space", "return"]) // "undo"
+
+    if currencyKeys.count < 3 {
+      return keyboardBuilder.build()
+    } else {
+      return keyboardBuilder
+        .replaceKey(row: 1, column: 6, to: currencyKeys[0])
+        .replaceKey(row: 1, column: 7, to: currencyKeys[1])
+        .replaceKey(row: 1, column: 8, to: currencyKeys[2])
+        .build()
+    }
+  }
+
+  // iPad keyboard layouts.
+  static func genPadLetterKeys() -> [[String]] {
+    return KeyboardBuilder()
+      .addRow(["1", "2", "3", "4", "5", "6", "7", "8", "9", "0", "-", "=", "+"])
+      .addRow(["q", "w", "e", "r", "t", "y", "u", "i", "o", "p", "delete"])
+      .addRow(["a", "s", "d", "f", "g", "h", "j", "k", "l", "ñ", "return"])
+      .addRow(["shift", "z", "x", "c", "v", "b", "n", "m", ",", ".", "shift"])
+      .addRow(["selectKeyboard", ".?123", "space", ".?123", "hideKeyboard"]) // "undo"
+      .build()
+  }
+
+  static func genPadDisableAccentsLetterKeys() -> [[String]] {
+    return KeyboardBuilder()
+      .addRow(["1", "2", "3", "4", "5", "6", "7", "8", "9", "0", "-", "=", "+"])
+      .addRow(["q", "w", "e", "r", "t", "y", "u", "i", "o", "p", "delete"])
+      .addRow(["a", "s", "d", "f", "g", "h", "j", "k", "l", "return"])
+      .addRow(["shift", "z", "x", "c", "v", "b", "n", "m", ",", ".", "shift"])
+      .addRow(["selectKeyboard", ".?123", "space", ".?123", "hideKeyboard"]) // "undo"
+      .build()
+  }
+
+  static func genPadNumberKeys(currencyKey: String) -> [[String]] {
+    return KeyboardBuilder()
+      .addRow(["1", "2", "3", "4", "5", "6", "7", "8", "9", "0", "delete"])
+      .addRow(["@", "#", "$", "&", "*", "(", ")", "'", "\"", "+", "return"])
+      .addRow(["#+=", "%", "_", "-", "=", "/", ";", ":", ",", ".", "#+="])
+      .addRow(["selectKeyboard", "ABC", "space", "ABC", "hideKeyboard"]) // "undo"
+      .replaceKey(row: 1, column: 2, to: currencyKey)
+      .build()
+  }
+
+  static func genPadSymbolKeys(currencyKeys: [String]) -> [[String]] {
+    let keyboardBuilder = KeyboardBuilder()
+      .addRow(["1", "2", "3", "4", "5", "6", "7", "8", "9", "0", "delete"])
+      .addRow(["€", "£", "¥", "^", "[", "]", "{", "}", "ᵒ", "ᵃ", "return"])
+      .addRow(["123", "§", "|", "~", "¶", "\\", "<", ">", "¡", "¿", "123"])
+      .addRow(["selectKeyboard", "ABC", "space", "ABC", "hideKeyboard"]) // "undo"
+
+    if currencyKeys.count < 3 {
+      return keyboardBuilder.build()
+    } else {
+      return keyboardBuilder
+        .replaceKey(row: 1, column: 0, to: currencyKeys[0])
+        .replaceKey(row: 1, column: 1, to: currencyKeys[1])
+        .replaceKey(row: 1, column: 2, to: currencyKeys[2])
+        .build()
+    }
+  }
+
+  // Expanded iPad keyboard layouts for wider devices.
+  static func genPadExpandedLetterKeys() -> [[String]] {
+    return KeyboardBuilder()
+    .addRow(["|", "1", "2", "3", "4", "5", "6", "7", "8", "9", "0", "?", "¿", "delete"])
+    .addRow([SpecialKeys.indent, "q", "w", "e", "r", "t", "y", "u", "i", "o", "p", "´", "+", "*"])
+    .addRow([SpecialKeys.capsLock, "a", "s", "d", "f", "g", "h", "j", "k", "l", "ñ", "{", "}", "return"])
+    .addRow(["shift", "'", "z", "x", "c", "v", "b", "n", "m", ",", ".", "-", "shift"])
+    .addRow(["selectKeyboard", ".?123", "space", ".?123", "hideKeyboard"]) // "microphone", "scribble"
+    .build()
+  }
+
+  static func genPadExpandedDisableAccentsLetterKeys() -> [[String]] {
+    return KeyboardBuilder()
+    .addRow(["|", "1", "2", "3", "4", "5", "6", "7", "8", "9", "0", "?", "¿", "delete"])
+    .addRow([SpecialKeys.indent, "q", "w", "e", "r", "t", "y", "u", "i", "o", "p", "´", "+", "*"])
+    .addRow([SpecialKeys.capsLock, "a", "s", "d", "f", "g", "h", "j", "k", "l", "~", "{", "}", "return"])
+    .addRow(["shift", "'", "z", "x", "c", "v", "b", "n", "m", ",", ".", "-", "shift"])
+    .addRow(["selectKeyboard", ".?123", "space", ".?123", "hideKeyboard"]) // "microphone", "scribble"
+    .build()
+}
+
+static func genPadExpandedSymbolKeys() -> [[String]] {
+  return KeyboardBuilder()
+    .addRow(["`", "1", "2", "3", "4", "5", "6", "7", "8", "9", "0", "<", ">", "delete"])
+    .addRow([SpecialKeys.indent, "(", ")", "{", "}", "#", "%", "^", "*", "+", "=", "\\", "|", "§"])
+    .addRow([SpecialKeys.capsLock, "—", "/", ":", ";", "&", "@", "$", "£", "¥", "~", "[", "]", "return"]) // "undo"
+    .addRow(["shift", "…", "?", "!", "≠", "'", "\"", "_", "€", ",", ".", "-", "shift"]) // "redo"
+    .addRow(["selectKeyboard", "ABC", "space", "ABC", "hideKeyboard"]) // "microphone", "scribble"
+    .build()
+}
+}
+
 /// Gets the keys for the Spanish keyboard.
 func getESKeys() {
-  let userDefaults = UserDefaults(suiteName: "group.be.scri.userDefaultsContainer")!
+  guard let userDefaults = UserDefaults(suiteName: "group.be.scri.userDefaultsContainer") else {
+    fatalError()
+  }
+
+  var currencyKey = SpanishKeyboardConstants.defaultCurrencyKey
+  var currencyKeys = SpanishKeyboardConstants.currencyKeys
+  let dictionaryKey = controllerLanguage + "defaultCurrencySymbol"
+  if let currencyValue = userDefaults.string(forKey: dictionaryKey) {
+    currencyKey = currencyValue
+  } else {
+    userDefaults.setValue(currencyKey, forKey: dictionaryKey)
+  }
+  if let index = currencyKeys.firstIndex(of: currencyKey) {
+    currencyKeys.remove(at: index)
+  }
 
   if DeviceType.isPhone {
     if userDefaults.bool(forKey: "esAccentCharacters") {
-      letterKeys = SpanishKeyboardConstants.letterKeysPhoneDisableAccents
+      letterKeys = SpanishKeyboardProvider.genPhoneDisableAccentsLetterKeys()
     } else {
-      letterKeys = SpanishKeyboardConstants.letterKeysPhone
+      letterKeys = SpanishKeyboardProvider.genPhoneLetterKeys()
     }
-    numberKeys = SpanishKeyboardConstants.numberKeysPhone
-    symbolKeys = SpanishKeyboardConstants.symbolKeysPhone
+    numberKeys = SpanishKeyboardProvider.genPhoneNumberKeys(currencyKey: currencyKey)
+    symbolKeys = SpanishKeyboardProvider.genPhoneSymbolKeys(currencyKeys: currencyKeys)
     allKeys = Array(letterKeys.joined()) + Array(numberKeys.joined()) + Array(symbolKeys.joined())
 
     leftKeyChars = ["q", "a", "1", "-", "[", "_"]
@@ -147,21 +208,21 @@ func getESKeys() {
     // Use the expanded keys layout if the iPad is wide enough and has no home button.
     if usingExpandedKeyboard {
       if userDefaults.bool(forKey: "esAccentCharacters") {
-        letterKeys = SpanishKeyboardConstants.letterKeysPadExpandedDisableAccents
+        letterKeys = SpanishKeyboardProvider.genPadExpandedDisableAccentsLetterKeys()
       } else {
-        letterKeys = SpanishKeyboardConstants.letterKeysPadExpanded
+        letterKeys = SpanishKeyboardProvider.genPadExpandedLetterKeys()
       }
-      symbolKeys = SpanishKeyboardConstants.symbolKeysPadExpanded
+      symbolKeys = SpanishKeyboardProvider.genPadExpandedSymbolKeys()
 
       allKeys = Array(letterKeys.joined()) + Array(symbolKeys.joined())
     } else {
       if userDefaults.bool(forKey: "esAccentCharacters") {
-        letterKeys = SpanishKeyboardConstants.letterKeysPadDisableAccents
+        letterKeys = SpanishKeyboardProvider.genPadDisableAccentsLetterKeys()
       } else {
-        letterKeys = SpanishKeyboardConstants.letterKeysPad
+        letterKeys = SpanishKeyboardProvider.genPadLetterKeys()
       }
-      numberKeys = SpanishKeyboardConstants.numberKeysPad
-      symbolKeys = SpanishKeyboardConstants.symbolKeysPad
+      numberKeys = SpanishKeyboardProvider.genPadNumberKeys(currencyKey: currencyKey)
+      symbolKeys = SpanishKeyboardProvider.genPadSymbolKeys(currencyKeys: currencyKeys)
 
       letterKeys.removeFirst(1)