+++ Scribe/Components/InfoChildTableViewCell/InfoChildTableViewCell.swift
@@ -151,6 +151,10 @@ final class InfoChildTableViewCell: UITableViewCell {
       let dictionaryKey = languageCode + "AccentCharacters"
       userDefaults.setValue(toggleSwitch.isOn, forKey: dictionaryKey)
 
+    case .doubleSpacePeriods:
+      let dictionaryKey = languageCode + "DoubleSpacePeriods"
+      userDefaults.setValue(toggleSwitch.isOn, forKey: dictionaryKey)
+
     case .autosuggestEmojis:
       let dictionaryKey = languageCode + "EmojiAutosuggest"
       userDefaults.setValue(toggleSwitch.isOn, forKey: dictionaryKey)
@@ -177,6 +181,14 @@ final class InfoChildTableViewCell: UITableViewCell {
         toggleSwitch.isOn = toggleValue
       } else {
         toggleSwitch.isOn = false  // Default value
+      }
+
+    case .doubleSpacePeriods:
+      let dictionaryKey = languageCode + "DoubleSpacePeriods"
+      if let toggleValue = userDefaults.object(forKey: dictionaryKey) as? Bool {
+        toggleSwitch.isOn = toggleValue
+      } else {
+        toggleSwitch.isOn = true  // Default value
       }
 
     case .autosuggestEmojis: