+++ Scribe/ParentTableCellModel.swift
@@ -77,6 +77,7 @@ enum SectionState: Equatable {
 
 enum UserInteractiveState {
   case toggleCommaAndPeriod
+  case doubleSpacePeriods
   case autosuggestEmojis
   case toggleAccentCharacters
   case none